Crate gio_unix

Rust Gio Unix bindings

Rust bindings and wrappers for gio, part of gtk-rs-core.

Gio 2.56 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.83.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
glib = "0.13"
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

See Also

glib

License

gio-unix is available under the MIT License, please refer to it.

Re-exports

pub use gio_unix_sys as ffi;

Modules

functions

prelude

Structs

DesktopAppInfo

GDesktopAppInfo is an implementation of [gio::AppInfo][crate::gio::AppInfo] based on desktop files.

FDMessage

This [gio::SocketControlMessage][crate::gio::SocketControlMessage] contains a [gio::UnixFDList][crate::gio::UnixFDList]. It may be sent using Gio::Socket::send_message() and received using Gio::Socket::receive_message() over UNIX sockets (ie: sockets in the G_SOCKET_FAMILY_UNIX family). The file descriptors are copied between processes by the kernel.

FileDescriptorBased

GFileDescriptorBased is an interface for file descriptor based IO.

InputStream

GUnixInputStream implements [gio::InputStream][crate::gio::InputStream] for reading from a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

MountEntryv2_84

Defines a Unix mount entry (e.g. /media/cdrom). This corresponds roughly to a mtab entry.

MountMonitor

Watches for changes to the set of mount entries and mount points in the system.

MountPoint

Defines a Unix mount point (e.g. /dev). This corresponds roughly to a fstab entry.

OutputStream

GUnixOutputStream implements [gio::OutputStream][crate::gio::OutputStream] for writing to a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)