gio_unix::prelude

Trait FilterInputStreamExt

pub trait FilterInputStreamExt: IsA<FilterInputStream> + 'static {
    // Provided methods
    fn base_stream(&self) -> InputStream { ... }
    fn closes_base_stream(&self) -> bool { ... }
    fn set_close_base_stream(&self, close_base: bool) { ... }
    fn connect_close_base_stream_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
}

Trait containing all FilterInputStream methods.

Implementors

BufferedInputStream, ConverterInputStream, FilterInputStream

Provided Methods

fn base_stream(&self) -> InputStream

Gets the base stream for the filter stream.

Returns

a #GInputStream.

fn closes_base_stream(&self) -> bool

Returns whether the base stream will be closed when @self is closed.

Returns

true if the base stream will be closed.

fn set_close_base_stream(&self, close_base: bool)

Sets whether the base stream will be closed when @self is closed.

close_base

true to close the base stream.

fn connect_close_base_stream_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.