gio_unix::prelude

Trait SocketConnectableExt

pub trait SocketConnectableExt: IsA<SocketConnectable> + 'static {
    // Provided methods
    fn enumerate(&self) -> SocketAddressEnumerator { ... }
    fn proxy_enumerate(&self) -> SocketAddressEnumerator { ... }
    fn to_string(&self) -> GString { ... }
}

Trait containing all SocketConnectable methods.

Implementors

InetSocketAddress, NativeSocketAddress, NetworkAddress, NetworkService, ProxyAddress, SocketAddress, SocketConnectable, UnixSocketAddress

Provided Methods

fn enumerate(&self) -> SocketAddressEnumerator

Creates a #GSocketAddressEnumerator for @self.

Returns

a new #GSocketAddressEnumerator.

fn proxy_enumerate(&self) -> SocketAddressEnumerator

Creates a #GSocketAddressEnumerator for @self that will return a #GProxyAddress for each of its addresses that you must connect to via a proxy.

If @self does not implement g_socket_connectable_proxy_enumerate(), this will fall back to calling g_socket_connectable_enumerate().

Returns

a new #GSocketAddressEnumerator.

fn to_string(&self) -> GString

Format a #GSocketConnectable as a string. This is a human-readable format for use in debugging output, and is not a stable serialization format. It is not suitable for use in user interfaces as it exposes too much information for a user.

If the #GSocketConnectable implementation does not support string formatting, the implementation’s type name will be returned as a fallback.

Returns

the formatted string

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.